+++ Backend/src/main/kotlin/com/luckyseven/backend/domain/settlement/dto/SettlementCreateRequest.kt
@@ -1,4 +1,4 @@
-package com.luckyseven.backend.domain.settlements.dto
+package com.luckyseven.backend.domain.settlement.dto
 
 import jakarta.validation.constraints.DecimalMin
 import jakarta.validation.constraints.Digits
