+++ Backend/src/test/java/com/luckyseven/backend/domain/settlements/app/SettlementServiceTest.java
@@ -12,12 +12,13 @@ import com.luckyseven.backend.domain.expense.entity.Expense;
 import com.luckyseven.backend.domain.expense.service.ExpenseService;
 import com.luckyseven.backend.domain.member.entity.Member;
 import com.luckyseven.backend.domain.member.service.MemberService;
-import com.luckyseven.backend.domain.settlements.dao.SettlementRepository;
-import com.luckyseven.backend.domain.settlements.dto.SettlementCreateRequest;
-import com.luckyseven.backend.domain.settlements.dto.SettlementResponse;
-import com.luckyseven.backend.domain.settlements.dto.SettlementSearchCondition;
-import com.luckyseven.backend.domain.settlements.dto.SettlementUpdateRequest;
-import com.luckyseven.backend.domain.settlements.entity.Settlement;
+import com.luckyseven.backend.domain.settlement.app.SettlementService;
+import com.luckyseven.backend.domain.settlement.dao.SettlementRepository;
+import com.luckyseven.backend.domain.settlement.dto.SettlementCreateRequest;
+import com.luckyseven.backend.domain.settlement.dto.SettlementResponse;
+import com.luckyseven.backend.domain.settlement.dto.SettlementSearchCondition;
+import com.luckyseven.backend.domain.settlement.dto.SettlementUpdateRequest;
+import com.luckyseven.backend.domain.settlement.entity.Settlement;
 import com.luckyseven.backend.domain.team.entity.Team;
 import com.luckyseven.backend.sharedkernel.exception.CustomLogicException;
 import com.luckyseven.backend.sharedkernel.exception.ExceptionCode;
