+++ Backend/src/test/java/com/luckyseven/backend/domain/settlements/dao/SettlementRepositoryTest.java
@@ -8,7 +8,9 @@ import com.luckyseven.backend.domain.expense.entity.Expense;
 import com.luckyseven.backend.domain.expense.enums.ExpenseCategory;
 import com.luckyseven.backend.domain.expense.enums.PaymentMethod;
 import com.luckyseven.backend.domain.member.entity.Member;
-import com.luckyseven.backend.domain.settlements.entity.Settlement;
+import com.luckyseven.backend.domain.settlement.dao.SettlementRepository;
+import com.luckyseven.backend.domain.settlement.dao.SettlementSpecification;
+import com.luckyseven.backend.domain.settlement.entity.Settlement;
 import com.luckyseven.backend.domain.team.entity.Team;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
